Create README.md
As of the beginning of 2022, foreclosures are on the rise. Coupled with the fact that home purchases tends to be the largest purchase a person makes in their lifetime, it can be unfortunate and even catastrophic to make a mistake when buying a home. While foreclosures and a volatile market may be concerning, prejudice is also a problem that can arise. Despite the existence of the fair-housing act, testimonials detail discriminatory practices that can occur for people of color when a formal appraiser is called. While foreclosures are often the worst-case scenario, the better case scenario of making prudent real estate decisions can help a household achieve financial stability and overall wealth. There is an abundance of real estate data, giving hope to these issues with the possibilities afforded through data science methods, such as machine learning.

Objectives:
1. Build a predictor of price that uses comparable sold properties and not price history/ formal appraisals
2. Implement a machine learning model to increase accuracy of predictor
3. Use external API's to extract data and use data wrangling methods to prepare dataset for analysis and data science methods
4. Expand analysis to most of Fairfax County to elucidate potential shortcomings of the model (ex. where linear relationship may not be sufficient)

Why Does this project exist: 
1. To design pragmatic real estate predictors based on recently sold properties
2. To provide insight that may help individuals in unfortunate situations, such as foreclosure
3. To expand the reach of data science into real estate

Given that homes tend to be the largest asset that people purchase in their lifetime, a lot of data is present on houses. While a lot of this data may be inaccessible, APIs can be used and sometimes large corporations (ex. Redfin, Realtor.com, etc.) make data available online. Regarding homebuyer decisions - while emotional connections and gut feelings can be important when purchasing a home, there are more objective factors relating to a home (lot size, home size, bedrooms, bathrooms, etc.) that may be important to consider in terms of home value. It can be speculated that if a family is ever in a financial pinch, the ability to sell a home quickly and for the highest price will provide relief and prevent devastating processes such as foreclosures.